CSS refactoring
