+++ package.json
@@ -16,6 +16,7 @@
     "react-router-dom": "^7.4.0",
     "react-scripts": "5.0.1",
     "redux": "^5.0.1",
+    "sass": "^1.86.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
