--- src/App.css
@@ -1,54 +1,0 @@
-.header {
-  background-color: tomato;
-  height: 70px;
-  color: #fff;
-  display: flex;
-  justify-content: space-around;
-  padding: 10px;
-  align-items: center;
-}
-.header input{
-  height: 30px;
-  width: 800px;
-}
-.prodactCard{
-  box-shadow: 3px 3px 3px 3px gray;
-  margin: 20px;
-  padding: 10px;
-  width: 250px;
-  height: 350px;
-}
-.CardPicture{
-  height: 200px;
-}
-.cardBlock{
-  display: flex;
-  flex-wrap: wrap;
-}
-.navBar{
-  box-shadow: 3px 3px 3px 3px rgb(128, 128, 128);
-  width: 300px;
-  height: 100%;
-  position: absolute;
-  background: rgb(238, 238, 238);
-  opacity: 0.85;
-  padding: 10px;
-}
-.navBar div{
-  padding: 20px;
-}
-.navBar div:hover, .active{
-  background-color: rgb(57, 57, 182);
-}
-.cardContent{
-  display: flex;
-  justify-content: space-between;
-}
-.iconFavorite{
-  margin-right: 20px;
-  margin-bottom: 50px;
-  cursor: pointer;
-}
-.favoriteMenu{
-  margin-right: 45px;
-}
