+++ src/App.scss
@@ -1,0 +1,4 @@
+.cardBlock{
+  display: flex;
+  flex-wrap: wrap;
+}
