+++ src/Components/Sort/sort.scss
@@ -2,13 +2,14 @@
 	margin-top: 30px;
 	margin-left: 40px;
 	margin-bottom: 10px;
-}
-.sort span:first-child {
-	cursor:default;
-}
-.sort span {
-	margin-right: 20px;
-	cursor: pointer;
+	span:first-child {
+		cursor:default;
+	}
+
+	span {
+		margin-right: 20px;
+		cursor: pointer;
+	}
 }
 .sortActive{
 color: rgb(55, 55, 255);
