+++ src/Components/productCard/index.scss
@@ -1,0 +1,32 @@
+@mixin space-between {
+  display: flex;
+  justify-content: space-between;
+}
+
+.prodactCard {
+  box-shadow: 3px 3px 3px 3px gray;
+  margin: 20px;
+  padding: 10px;
+  width: 250px;
+  height: 350px;
+}
+
+.CardPicture {
+  height: 200px;
+}
+
+.cardBlock {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.cardContent {
+  @include space-between;
+}
+
+.iconsMainBlock {
+  @include space-between;
+  flex-direction: column;
+  margin-right: 25px;
+  margin-bottom: 20px;
+}
