+++ src/pages/cart/index.scss
@@ -1,3 +1,5 @@
+
+
 .cartItemBlock{
 	box-shadow: 1px 1px 1px 1px gray;
   margin: 20px;
@@ -16,10 +18,12 @@
 }
 .cartItemQuantity{
 	margin-right: 150px;
+
+	span{
+		margin: 0 20px;
+	}
 }
-.cartItemQuantity span{
-	margin: 0 20px;
-}
+
 .cartItemPrice{
 	width: 100px;
 }
