--- src/pages/product/index.css
@@ -1,9 +1,0 @@
-.productPageBlock{
-	display: flex;
-	box-shadow: 3px 3px 3px 3px gray;
-  margin: 20px;
-  padding: 10px;
-}
-.productInfo{
-	margin-left: 70px;
-}
