+++ src/pages/product/index.jsx
@@ -2,7 +2,7 @@ import { useParams } from "react-router-dom";
 import { loadProduct } from "./slices";
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
-import "./index.css";
+import "./index.scss";
 import { ToCartButton } from "../../Components/toCartButton";
 import { ToFavoriteButton } from "../../Components/toFavoriteButton";
 
@@ -22,7 +22,7 @@ export const Product = () => {
     return <div>Loading...</div>;
   }
 
-  const { img, name, brand, rating, price } = product;
+  const { img, name, brand, rating, price,description } = product;
 
   return (
     <div className="productPageBlock">
@@ -33,9 +33,10 @@ export const Product = () => {
           <h3>{brand}</h3>
           <div>рейтинг:{rating}</div>
           <h3>${price}</h3>
+          <p>{description}</p>
         </div>
 
-        <div>
+        <div className="productPageIcons">
           <ToFavoriteButton product={product}/>
 					<ToCartButton product={product}/>
         </div>
